Address review: move URL coherence tests into title_bar browsertest
- Fold FocusModeUrlCoherenceBrowserTest into the existing
  focus_mode_title_bar_view_browsertest.cc as a subclass of
  FocusModeTitleBarViewBrowserTest (reuses helpers, no fixture copy).
- Delete the standalone focus_mode_url_coherence_browsertest.cc and its
  BUILD.gn entry.
- Enable back/forward cache via
  content::GetDefaultEnabledBackForwardCacheFeaturesForTesting() and
  assert the prior RenderFrameHost is in kInBackForwardCache before
  GoBack, so the restore test exercises the cached-RFH path.
- Add #include content/public/browser/navigation_controller.h (IWYU for
  GetController().GoBack()).
- Drop the prerender edge-case comment block (authoring notes).
- Append << base::UTF16ToUTF8(domain_text()) to WaitForDomainToContain
  assertions in the moved tests so timeouts print the current label.
