+++ browser/ui/views/frame/BUILD.gn
@@ -98,7 +98,6 @@ source_set("browser_tests") {
     "//brave/browser/ui/views/frame/focus_mode_mini_toolbar_browsertest.cc",
     "//brave/browser/ui/views/frame/focus_mode_title_bar_view_browsertest.cc",
     "//brave/browser/ui/views/frame/focus_mode_top_overlay_browsertest.cc",
-    "//brave/browser/ui/views/frame/focus_mode_url_coherence_browsertest.cc",
     "//brave/browser/ui/views/frame/tab_strip_placement_coordinator_browsertest.cc",
   ]
